Model changes and rebuilding the drive nodes from the ground up
A number of problems were found with the drivenode.py implementation.  I decided to start
from the beginning writing a drive node based on arbotix.  I was able to get a simulated
node to move the model in rviz.  From there, I was able to add basic message handling from
the halnode and to see the real arlobot visualized in rviz.  Additionally, there were
problems found in the Psoc code where the angular velocity was being used without first
being convered to meters/second.  This showed up when attempt to do isolated moves, e.g.,
moving in a straight line.  In which case the linear velocity of the wheels was not the
same as the linear velocity of the unicycle model -- an impossibility if we're to get
accurate motion.  Anyway, the psoc code is fixed and now I need to build up the drive
node and the rest of the arlobot model.