Fixed imu visualization
Added an imu module to the toplevel urdf.  It was necessary to create
a link from the IMU to base_link (the imu sits on top of the robot).
Also had to update imustate so that it would use the imu link as the
frame id.
Also documented the xboxdrv installation